refactor(focus): extract storage helpers, slim coordinator class
Split src/lib/focus/storage.ts (633 LOC) into:
- storage.ts (209 LOC) - thin read-mutate-write coordinator
- storage-transitions.ts (263 LOC) - state machine transitions
- storage-mutations.ts (119 LOC) - add/remove/read helpers

Public API of focusStore singleton preserved; all 99 focus tests
plus 1842/1842 full suite pass. Removes storage.ts from size budget
baseline (5 production files remaining).
